# src/main/resources/static/swagger-api/swagger.yaml
...
openapi: 3.0.0
info:
  version: "1.0.0"
  title: "st-schools"
servers:
  - url: "http://localhost:8080"
...
paths:
...
  /api/v1/auth/login:
    post:
      tags:
        - "AuthenticationController"
      summary: "Login"
      operationId: "login"
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: "thongchuthanh200@gmail.com"
                password:
                  type: string
                  example: "1234567890"
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/AuthenticationResponse"
        "403":
          description: "Forbidden"
          content:
            'application/json':
              schema:
                type: object
                properties:
                  data:
                    type: string
                    example: "Incorrect password or email"
  /api/v1/auth/forgot:
    post:
      tags:
        - "AuthenticationController"
      summary: "Send password reset code to users email"
      operationId: "forgotPassword"
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: "test123@test.com"
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: object
                properties:
                  data:
                    type: string
                    example: "Reset password code is send to your E-mail"
        "400":
          description: "Bad Request"
          content:
            'application/json':
              schema:
                type: object
                properties:
                  data:
                    type: string
                    example: "Email not found"
  /api/v1/auth/reset/{code}:
    get:
      tags:
        - "AuthenticationController"
      summary: "Returns user info by password reset code"
      operationId: "getPasswordResetCode"
      parameters:
        - name: "code"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: "Bad Request"
          content:
            'application/json':
              schema:
                type: object
                properties:
                  data:
                    type: string
                    example: "Password reset code is invalid!"
  /api/v1/auth/reset:
    post:
      tags:
        - "AuthenticationController"
      summary: "Reset password"
      operationId: "passwordReset"
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              type: object
              properties:
                password:
                  type: string
                  example: "1234567890"
                password2:
                  type: string
                  example: "1234567890"
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: object
                properties:
                  data:
                    type: string
                    example: "Password successfully changed!"
        "400":
          description: "Bad Request"
          content:
            'application/json':
              schema:
                type: object
                properties:
                  passwordError:
                    type: "string"
                  password2Error:
                    type: "string"
              examples:
                PasswordMatchError:
                  summary: "Password Match Error"
                  value:
                    passwordError: "Passwords do not match."
                PasswordConfirmationError:
                  summary: "Password Confirmation Error"
                  value:
                    password2Error: "Password confirmation cannot be empty."

#  /api/v1/perfumes:
#    get:
#      tags:
#        - "PerfumeController"
#      summary: "Returns list of perfumes"
#      operationId: "getAllPerfumes"
#      parameters: []
#      responses:
#        "200":
#          description: "OK"
#          content:
#            'application/json':
#              schema:
#                type: array
#                items:
#                  $ref: "#/components/schemas/PerfumeResponse"
#
#  /api/v1/perfumes/{id}:
#    get:
#      tags:
#        - "PerfumeController"
#      summary: "Get perfume by id"
#      operationId: "getPerfume"
#      parameters:
#        - name: "id"
#          in: "path"
#          required: true
#          schema:
#            type: "integer"
#            format: "int64"
#      responses:
#        "200":
#          description: "OK"
#          content:
#            'application/json':
#              schema:
#                $ref: "#/components/schemas/PerfumeResponse"
#
#  /api/v1/perfumes/ids:
#    post:
#      tags:
#        - "PerfumeController"
#      summary: "Get perfumes by ids"
#      operationId: "getPerfumesByIds"
#      requestBody:
#        content:
#          'application/json':
#            schema:
#              type: "array"
#              items:
#                type: "integer"
#                format: int64
#              example: [1, 2, 3]
#      responses:
#        "200":
#          description: "OK"
#          content:
#            'application/json':
#              schema:
#                type: array
#                items:
#                  $ref: "#/components/schemas/PerfumeResponse"
#
#  /api/v1/perfumes/search:
#    post:
#      tags:
#        - "PerfumeController"
#      summary: "Find perfumes by filter parameters"
#      operationId: "findPerfumesByFilterParams"
#      requestBody:
#        required: true
#        content:
#          'application/json':
#            schema:
#              type: object
#              properties:
#                perfumers:
#                  type: "array"
#                  items:
#                    type: "string"
#                    example: "Dior"
#                genders:
#                  type: "array"
#                  items:
#                    type: "string"
#                    example: "male"
#                prices:
#                  type: "array"
#                  items:
#                    type: "integer"
#                    format: int32
#                  example: [1, 200]
#      responses:
#        "200":
#          description: "OK"
#          content:
#            'application/json':
#              schema:
#                type: array
#                items:
#                  $ref: "#/components/schemas/PerfumeResponse"
#
#  /api/v1/perfumes/search/gender:
#    post:
#      tags:
#        - "PerfumeController"
#      summary: "Find perfumes by gender"
#      operationId: "findByPerfumeGender"
#      requestBody:
#        required: true
#        content:
#          'application/json':
#            schema:
#              type: object
#              properties:
#                genders:
#                  type: "array"
#                  items:
#                    type: "string"
#                    example: "male"
#      responses:
#        "200":
#          description: "OK"
#          content:
#            'application/json':
#              schema:
#                type: array
#                items:
#                  $ref: "#/components/schemas/PerfumeResponse"
#
#  /api/v1/perfumes/search/perfumer:
#    post:
#      tags:
#        - "PerfumeController"
#      summary: "Find perfumes by brand"
#      operationId: "findByPerfumer"
#      requestBody:
#        required: true
#        content:
#          'application/json':
#            schema:
#              type: object
#              properties:
#                perfumers:
#                  type: "array"
#                  items:
#                    type: "string"
#                  example: ["Dior", "Creed"]
#      responses:
#        "200":
#          description: "OK"
#          content:
#            'application/json':
#              schema:
#                type: array
#                items:
#                  $ref: "#/components/schemas/PerfumeResponse"
#
#  /api/v1/perfumes/graphql/ids:
#    post:
#      tags:
#        - "PerfumeController"
#      summary: "Get list of perfumes by ids GraphQL query"
#      operationId: "getPerfumesByIdsQuery"
#      requestBody:
#        required: true
#        content:
#          'application/json':
#            schema:
#              type: object
#              properties:
#                query:
#                  type: "string"
#                  example: "{ perfumesIds(ids: [16,17,18]) {id perfumeTitle perfumer price filename} }"
#
#      responses:
#        "200":
#          description: "OK"
#          content:
#            'application/json':
#              schema:
#                type: array
#                items:
#                  type: object
#                  properties:
#                    id:
#                      type: "integer"
#                      format: int64
#                    perfumeTitle:
#                      type: "string"
#                    perfumer:
#                      type: "string"
#                    price:
#                      type: "integer"
#                      format: int32
#                    filename:
#                      type: "string"
#                  example:
#                    id: 43
#                    perfumeTitle: "Sauvage"
#                    perfumer: "Dior"
#                    price: 62
#                    filename: "empty.jpg"
#
#  /api/v1/perfumes/graphql/perfumes:
#    post:
#      tags:
#        - "PerfumeController"
#      summary: "Get list of perfumes by GraphQL query"
#      operationId: "getAllPerfumesByQuery"
#      requestBody:
#        required: true
#        content:
#          'application/json':
#            schema:
#              type: object
#              properties:
#                query:
#                  type: "string"
#                  example: "{ perfumes {id perfumeTitle perfumer price filename} }"
#
#      responses:
#        "200":
#          description: "OK"
#          content:
#            'application/json':
#              schema:
#                type: array
#                items:
#                  type: object
#                  properties:
#                    id:
#                      type: "integer"
#                      format: int64
#                    perfumeTitle:
#                      type: "string"
#                    perfumer:
#                      type: "string"
#                    price:
#                      type: "integer"
#                      format: int32
#                    filename:
#                      type: "string"
#                  example:
#                    id: 43
#                    perfumeTitle: "Sauvage"
#                    perfumer: "Dior"
#                    price: 62
#                    filename: "empty.jpg"
#
#  /api/v1/perfumes/graphql/perfume:
#    post:
#      tags:
#        - "PerfumeController"
#      summary: "Get perfume by GraphQL query"
#      operationId: "getPerfumeByQuery"
#      requestBody:
#        required: true
#        content:
#          'application/json':
#            schema:
#              type: object
#              properties:
#                query:
#                  type: "string"
#                  example: "{ perfume(id: 43) {id perfumeTitle perfumer price filename} }"
#      responses:
#        "200":
#          description: "OK"
#          content:
#            'application/json':
#              schema:
#                type: object
#                properties:
#                  id:
#                    type: "integer"
#                    format: int64
#                  perfumeTitle:
#                    type: "string"
#                  perfumer:
#                    type: "string"
#                  price:
#                    type: "integer"
#                    format: int32
#                  filename:
#                    type: "string"
#                example:
#                  id: 43
#                  perfumeTitle: "Sauvage"
#                  perfumer: "Dior"
#                  price: 62
#                  filename: "empty.jpg"
...
  /api/v1/users/edit/password:
    put:
      tags:
        - "UserController"
      summary: "Change user password"
      operationId: "updateUserPassword"
      requestBody:
        content:
          'application/json':
            schema:
              type: object
              properties:
                password:
                  type: "string"
                  example: "1234567890"
                password2:
                  type: "string"
                  example: "1234567890"
              required:
                - password
                - password2
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: object
                properties:
                  data:
                    type: string
                    example: "Password successfully changed!"
        "400":
          description: "Bad Request"
          content:
            'application/json':
              schema:
                type: object
                properties:
                  passwordError:
                    type: "string"
                  password2Error:
                    type: "string"
              examples:
                BindingResultError:
                  summary: "Binding Result Errors"
                  value:
                    passwordError: "The password must be between 6 and 16 characters long"
                    password2Error: "The password confirmation must be between 6 and 16 characters long"
                MatchPasswordError:
                  summary: "Match Password Error"
                  value:
                    data: "Passwords do not match."
      security:
        - token: []

#  /api/v1/users/cart:
#    post:
#      tags:
#        - "UserController"
#      summary: "Get perfumes by ids"
#      operationId: "getCart"
#      requestBody:
#        content:
#          'application/json':
#            schema:
#              type: "array"
#              items:
#                type: "integer"
#                format: int32
#              example: [1, 2, 3]
#      responses:
#        "200":
#          description: "OK"
#          content:
#            'application/json':
#              schema:
#                type: array
#                items:
#                  $ref: "#/components/schemas/PerfumeResponse"

#  /api/v1/users/orders:
#    get:
#      tags:
#        - "UserController"
#      summary: "Get all user orders by email"
#      operationId: "getAllUserOrders"
#      parameters: []
#      responses:
#        "200":
#          description: "OK"
#          content:
#            'application/json':
#              schema:
#                type: array
#                items:
#                  $ref: "#/components/schemas/OrderResponse"
#      security:
#        - token: []

#  /api/v1/user/graphql/orders:
#    post:
#      tags:
#        - "UserController"
#      summary: "Get all user orders by GraphQL query"
#      operationId: "getUserOrdersByQuery"
#      requestBody:
#        required: true
#        content:
#          'application/json':
#            schema:
#              type: object
#              properties:
#                query:
#                  type: "string"
#                  example: "{ ordersByEmail(email: test123@test.com) { id totalPrice date firstName lastName city address email phoneNumber postIndex orderItems { id amount quantity perfume { id perfumeTitle perfumer price filename } } } }"
#      responses:
#        "200":
#          description: "OK"
#          content:
#            'application/json':
#              schema:
#                type: array
#                items:
#                  $ref: "#/components/schemas/OrderResponse"
#      security:
#        - token: []


#  /api/v1/users/order:
#    post:
#      tags:
#        - "UserController"
#      summary: "Post order"
#      operationId: "postOrder"
#      requestBody:
#        content:
#          'application/json':
#            schema:
#              type: object
#              properties:
#                totalPrice:
#                  type: "number"
#                  format: "double"
#                  example: 100
#                firstName:
#                  type: "string"
#                  example: "John"
#                lastName:
#                  type: "string"
#                  example: "Doe"
#                city:
#                  type: "string"
#                  example: "New York"
#                address:
#                  type: "string"
#                  example: "Wall Street 1"
#                email:
#                  type: "string"
#                  example: "test123@test.com"
#                phoneNumber:
#                  type: "string"
#                  example: "123456"
#                postIndex:
#                  type: "integer"
#                  format: "int32"
#                  example: 123456
#                perfumesId:
#                  type: "object"
#                  properties:
#                    43:
#                      type: integer
#                      example: 1
#              required:
#                - firstName
#                - lastName
#                - city
#                - address
#                - email
#                - phoneNumber
#                - postIndex
#      responses:
#        "200":
#          description: "OK"
#          content:
#            'application/json':
#              schema:
#                type: array
#                items:
#                  $ref: "#/components/schemas/OrderResponse"
#        "400":
#          description: "Bad Request"
#          content:
#            'application/json':
#              schema:
#                type: object
#                properties:
#                  firstNameError:
#                    type: "string"
#                    example: "Fill in the input field"
#                  lastNameError:
#                    type: "string"
#                    example: "Fill in the input field"
#                  cityError:
#                    type: "string"
#                    example: "Fill in the input field"
#                  addressError:
#                    type: "string"
#                    example: "Fill in the input field"
#                  emailError:
#                    type: "string"
#                    example: "Email cannot be empty"
#                  phoneNumberError:
#                    type: "string"
#                    example: "Phone number cannot be empty"
#                  postIndexError:
#                    type: "string"
#                    example: "Post index cannot be empty"

#  /api/v1/users/review:
#    post:
#      tags:
#        - "UserController"
#      summary: "Add review to perfume"
#      operationId: "addReviewToPerfume"
#      requestBody:
#        content:
#          'application/json':
#            schema:
#              type: object
#              properties:
#                perfumeId:
#                  type: "number"
#                  format: "int64"
#                  example: 43
#                author:
#                  type: "string"
#                  example: "John Doe"
#                message:
#                  type: "string"
#                  example: "Hello World!"
#              required:
#                - perfumeId
#                - author
#                - message
#      responses:
#        "200":
#          description: "OK"
#          content:
#            'application/json':
#              schema:
#                type: object
#                properties:
#                  data:
#                    type: string
#                    example: "Review added successfully."
#        "400":
#          description: "Bad Request"
#          content:
#            'application/json':
#              schema:
#                type: object
#                properties:
#                  authorError:
#                    type: "string"
#                    example: "Fill in the input field"
#                  messageError:
#                    type: "string"
#                    example: "Fill in the input field"
components:
  securitySchemes:
    token:
      type: apiKey
      name: Authorization
      in: header
  schemas:
...
    User:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          example: 1
        email:
          type: "string"
          example: "thongchuthanh2000@test.com"
        firstName:
          type: "string"
          example: "Cheng"
        lastName:
          type: "string"
          example: "Tang Yu"
        avatar:
          type: "string"
          example: ""
        address:
          type: "string"
          example: "BR-VT"
        phone:
          type: "string"
          example: "0918948074"
        provider:
          type: "string"
          example: "LOCAL"
        active:
          type: "boolean"
          example: true
        activationCode:
          type: "string"
          example: ""
        passwordResetCode:
          type: "string"
          example: ""
        roles:
          type: "array"
          items:
            type: "string"
            example: "USER"
    AuthenticationResponse:
      type: "object"
      properties:
        email:
          type: "string"
        token:
          type: "string"
        userRole:
          type: "string"